status api  added

+++ src/main/java/com/example/demo/repository/StaffMemberRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.web.models.StaffMember;
 
-public interface StaffMemberRepository extends JpaRepository<StaffMember, Integer> {
+public interface StaffMemberRepository extends JpaRepository<StaffMember, Long> {
 }
